Renamed project and all references to `tags`
