+++ bower.json
@@ -1,5 +1,5 @@
 {
-  "name": "ember-cli-deploy-tag",
+  "name": "ember-cli-deploy-revision-key",
   "dependencies": {
     "ember": "1.11.1",
     "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
