+++ lib/key-generators/file-hash.js
