+++ lib/key-generators/index.js
